client.js を main.js
